backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlmodel
pydantic
sqlalchemy
python-decouple
psycopg2
PyJWT
passlib